pyproject.toml
[project]
name = "forbin"
version = "0.1.0"
description = "Interactive CLI tool for testing remote MCP servers - inspired by Colossus: The Forbin Project, where systems learn to communicate"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "fastapi",
    "fastmcp",
    "testing",
    "cli",
    "agentic",
    "ai-tools",
    "forbin"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
]


[project.urls]
Repository = "https://github.com/chris-colinsky/Forbin"
Issues = "https://github.com/chris-colinsky/Forbin/issues"
Documentation = "https://github.com/chris-colinsky/Forbin/blob/main/README.md"

[project.scripts]
forbin = "forbin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forbin"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
